+++ branch-misprediction/branch-target-misprediction.cpp
@@ -37,7 +37,7 @@ size_t test_memory(const std::vector<A*>& memory)
 
     for (int i = 0; i < REPETITIONS; i++)
     {
-        for (int j = 0; j < size; j++)
+        for (size_t j = 0; j < size; j++)
         {
             memory[j]->handle(&sum);
         }
